requirements.txt
Flask
Flask-OAuthlib
google-auth
google-auth-oauthlib
Werkzeug>=2.2,<3.0
pillow
